+++ src/core/types/type_float.h
@@ -21,7 +21,7 @@
 //------------------------------------------------------------------------------
 #ifndef dt_TYPES_TYPE_FLOAT_h
 #define dt_TYPES_TYPE_FLOAT_h
-#include "types/type_numeric.h"
+#include "types/typeimpl_numeric.h"
 namespace dt {
