+++ src/core/types/type_string.h
@@ -22,7 +22,7 @@
 #ifndef dt_TYPES_TYPE_STRING_h
 #define dt_TYPES_TYPE_STRING_h
 #include "python/string.h"
-#include "types/type_impl.h"
+#include "types/typeimpl.h"
 #include "types/type_invalid.h"
 #include "utils/assert.h"
 namespace dt {
